test(style): remove pydocstyle tests on charmcraft code
Also moves the functions only used when testing the init command to the
appropriate file.

We're removing these because they're duplicated by our use of ruff.
